Add copyright notices. Refs #1.
This commit adjust the copyright header to the cabal file, to the haskell file
and the LICENSE file.

The copyright year is kept as 2011, because the haskell file had a note
with that exact year. It was adapted to match the format expected by
haddock.

The cabal file is documented, to match.

The license file adds Keera Studios as credited for all changes after
2013, looking at the repo log to find the first change.

The support email address is changed to Keera's.